moved  nav logic and functionality from content editor into content nav, clean up
