+++ src/apps/content-editor/src/app/ContentEditor.less
@@ -6,19 +6,6 @@
   grid-template-columns: 60px 1fr;
   position: relative;
 
-  .Nav {
-    overflow: hidden;
-    box-shadow: none;
-
-    //Hide Scrollbar
-    scrollbar-width: none; /*  FireFox */
-    -ms-overflow-style: none; /*  Edge */
-    ::-webkit-scrollbar {
-      /*  Chrome / Safari */
-      display: none;
-    }
-  }
-
   .Content {
     overflow: hidden;
     box-shadow: -1px 1px 10px 2px fade(@zesty-dark-blue, 50%);
@@ -31,46 +18,16 @@
     }
   }
 
-  &.ContentNavHover,
-  &.ContentNavOpen {
+  &.ContentGridHover,
+  &.ContentGridOpen {
     grid-template-columns: 20vw auto;
 
     @media only screen and (min-width: 2000px) {
       grid-template-columns: 15vw auto;
     }
-
-    .Nav {
-      overflow-y: auto;
-      overflow-x: hidden;
-      height: calc(100vh - 54px);
-      &::-webkit-scrollbar {
-        width: 0;
-        height: 0;
-      }
-    }
   }
 
-  &.ContentNavHover {
-    .Nav {
-      background-color: @appBkgColor;
-      box-shadow: -1px 1px 10px 2px fade(@zesty-dark-blue, 50%);
-
-      position: absolute;
-      left: 0;
-      top: 0;
-      z-index: 10;
-      width: 20vw;
-
-      ::-webkit-scrollbar {
-        /*  Chrome / Safari */
-        display: none;
-      }
-
-      @media screen and (max-width: 1680px) {
-        width: 25vw;
-      }
-    }
-
+  &.ContentGridHover {
     .Content {
       position: absolute;
       top: 0;
@@ -85,14 +42,6 @@
       &.GlobalClosed {
         // 100vw - GlobalSidebar - contentNavCollapsed
         width: calc(100vw - 60px - 60px);
-      }
-    }
-  }
-  &.ContentNavOpen {
-    .Nav {
-      width: 20vw;
-      @media only screen and (min-width: 2000px) {
-        width: 15vw;
       }
     }
   }
