+++ src/apps/content-editor/src/app/components/Nav/ContentNav.less
@@ -1,5 +1,53 @@
 @import "~@zesty-io/core/colors.less";
 
+.NavContainer {
+  overflow: hidden;
+  box-shadow: none;
+  height: calc(100vh - 54px);
+
+  //Hide Scrollbar
+  scrollbar-width: none; /*  FireFox */
+  -ms-overflow-style: none; /*  Edge */
+  ::-webkit-scrollbar {
+    /*  Chrome / Safari */
+    display: none;
+  }
+
+  &.ContentNavHover,
+  &.ContentNavOpen {
+    background-color: @appBkgColor;
+
+    overflow-y: auto;
+    overflow-x: hidden;
+    height: calc(100vh - 54px);
+    &::-webkit-scrollbar {
+      width: 0;
+      height: 0;
+      display: none;
+    }
+  }
+
+  &.ContentNavHover {
+    box-shadow: -1px 1px 10px 2px fade(@zesty-dark-blue, 50%);
+
+    position: absolute;
+    left: 0;
+    top: 0;
+    z-index: 10;
+    width: 20vw;
+
+    @media screen and (max-width: 1680px) {
+      width: 25vw;
+    }
+  }
+  &.ContentNavOpen {
+    width: 20vw;
+    @media only screen and (min-width: 2000px) {
+      width: 15vw;
+    }
+  }
+}
+
 .SearchModels {
   margin: 8px 8px 0;
   input {
