+++ src/apps/content-editor/src/app/components/Nav/components/ContentNavToggle/ContentNavToggle.less
@@ -20,7 +20,9 @@
   }
 
   &.ContentNavToggleOpen {
+    z-index: 10;
     width: 20vw;
+
     @media only screen and (min-width: 2000px) {
       width: 15vw;
     }
